+++ demos/CommandLine/Models/Supabase/List.cs
@@ -1,4 +1,3 @@
-
 using Newtonsoft.Json;
 
 using Supabase.Postgrest.Attributes;